tests added

+++ x/vendorinfo/genesis_test.go
@@ -28,7 +28,7 @@ func (m *DclauthKeeperMock) HasRole(
 func (m *DclauthKeeperMock) HasVendorID(
 	ctx sdk.Context,
 	addr sdk.AccAddress,
-	vid uint64,
+	vid int32,
 ) bool {
 	args := m.Called(ctx, addr, vid)
 	return args.Bool(0)
